requirements.txt
google-genai>=0.7.0
python-dotenv>=1.0.1